Re-organize tests to follow existing structure:
- mint_many functionality tests go into `test_mint_many`
- access tests go into `test_admin`
- removed some redundant testing remaining from v1 -> v2 test updates
- added test for owner can mint many
